requirements.txt
python-dotenv==1.0.0
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
requests==2.31.0
pandas==2.1.4
geopandas==0.14.1
pyodbc==4.0.39 